# REF: de-duplicate code for pad/backfill