--- a/pandas/core/arrays/string_arrow.py
+++ b/pandas/core/arrays/string_arrow.py
@@ -399,8 +399,9 @@ class ArrowStringArray(OpsMixin, ExtensionArray):
 
         if mask.any():
             if method is not None:
-                func = missing.get_fill_func(method)
-                new_values, _ = func(self.to_numpy(object), limit=limit, mask=mask)
+                new_values, _ = missing.interpolate_2d(
+                    self.to_numpy(object), method=method, limit=limit, mask=mask
+                )
                 new_values = self._from_sequence(new_values)
             else:
                 # fill with value