@jorisvandenbossche 2021-03-10
@simonjayhawkins does the slowdown come from the cython changes (removing the 1D version) or from other changes?